+++ pyproject.toml
@@ -1,5 +1,5 @@
 [project]
-name = "icl-x1"
+name = "infractl"
 version = "0.0.1"
 description = "API driven Machine Learning suite"
 license = {file = "LICENSE"}
@@ -27,7 +27,7 @@ dependencies = [
 ]
 
 [project.scripts]
-x1 = "x1.cli.main:main"
+infractl = "infractl.cli.main:main"
 
 [project.optional-dependencies]
 build = [
@@ -61,8 +61,8 @@ build-backend = "setuptools.build_meta"
 
 [tool.setuptools.packages.find]
 where = ["src"]
-include = ["x1*"]
-exclude = ["x1.tests.*"]
+include = ["infractl*"]
+exclude = ["infractl.tests.*"]
 namespaces = false
 
 [tool.black]
@@ -82,11 +82,11 @@ color = true
 # TODO: increase percentage gradually
 fail-under = 62
 # Exclude tests
-exclude = ["src/x1/tests", "scripts/tests", "tests", "build"]
+exclude = ["src/infractl/tests", "scripts/tests", "tests", "build"]
 
 [tool.bandit]
 # ignore unit and integration tests, since they use "assert"
-exclude_dirs = ["src/x1/tests", "scripts/tests", "tests", "build"]
+exclude_dirs = ["src/infractl/tests", "scripts/tests", "tests", "build"]
 # tests = []
 skips = [
     "B404", # consider possible security implications associated with the subprocess module.
