+++ src/infractl/base/registry.py
