+++ src/infractl/tests/base/test_registry.py
