--- tests/integration/__init__.py
@@ -1,1 +1,0 @@
-"""ICL on X1 cluster integration tests."""
